Make comments user and location specific

+++ app/api/comments/[commentId]/route.ts
@@ -1,4 +1,6 @@
+import { cookies } from 'next/headers';
 import { NextRequest, NextResponse } from 'next/server';
+import { comment } from 'postcss';
 import { z } from 'zod';
 import {
   Comment,
@@ -6,6 +8,7 @@ import {
   getCommentById,
   updateCommentById,
 } from '../../../../database/comments';
+import { getUserBySessionToken } from '../../../../database/users';
 
 const commentType = z.object({
   content: z.string(),
@@ -47,7 +50,21 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: Record<string, string | string[]> },
 ): Promise<NextResponse<CommentResponseBodyDelete>> {
+  // this is a protected Route Handler
+  // 1. get the session token from the cookie
+  const cookieStore = cookies();
+  const token = cookieStore.get('sessionToken');
+
+  // 2. validate that session
+  // 3. get the user profile matching the session
+  const user = token && (await getUserBySessionToken(token.value));
+
+  if (!user) {
+    return NextResponse.json({ error: 'session token is not valid' });
+  }
+
   const commentId = Number(params.commentId);
+  const singleCommentCheck = await getCommentById(commentId);
 
   if (!commentId) {
     return NextResponse.json(
@@ -58,20 +75,41 @@ export async function DELETE(
     );
   }
 
-  const singleComment = await deleteCommentById(commentId);
+  if (singleCommentCheck && singleCommentCheck.userId === user.id) {
+    const singleComment = await deleteCommentById(commentId);
 
-  if (!singleComment) {
-    return NextResponse.json({ error: 'Comment not found' }, { status: 404 });
+    if (!singleComment) {
+      return NextResponse.json({ error: 'Comment not found' }, { status: 404 });
+    }
+
+    return NextResponse.json({ comment: singleComment });
+  } else {
+    return NextResponse.json(
+      { error: 'You cannot delete this comment, you sausage!!!' },
+      { status: 404 },
+    );
   }
-
-  return NextResponse.json({ comment: singleComment });
 }
 
 export async function PUT(
   request: NextRequest,
   { params }: { params: Record<string, string | string[]> },
 ): Promise<NextResponse<CommentResponseBodyPut>> {
+  // this is a protected Route Handler
+  // 1. get the session token from the cookie
+  const cookieStore = cookies();
+  const token = cookieStore.get('sessionToken');
+
+  // 2. validate that session
+  // 3. get the user profile matching the session
+  const user = token && (await getUserBySessionToken(token.value));
+
+  if (!user) {
+    return NextResponse.json({ error: 'session token is not valid' });
+  }
+
   const commentId = Number(params.commentId);
+  const singleCommentCheck = await getCommentById(commentId);
 
   if (!commentId) {
     return NextResponse.json(
@@ -95,11 +133,18 @@ export async function PUT(
     );
   }
 
-  const newComment = await updateCommentById(commentId, result.data.content);
+  if (singleCommentCheck && singleCommentCheck.userId === user.id) {
+    const newComment = await updateCommentById(commentId, result.data.content);
 
-  if (!newComment) {
-    return NextResponse.json({ error: 'Comment not found' }, { status: 404 });
+    if (!newComment) {
+      return NextResponse.json({ error: 'Comment not found' }, { status: 404 });
+    }
+
+    return NextResponse.json({ comment: newComment });
+  } else {
+    return NextResponse.json(
+      { error: 'You cannot edit this comment, you sausage!!!' },
+      { status: 404 },
+    );
   }
-
-  return NextResponse.json({ comment: newComment });
 }
